+++ src/proxy/seat.rs
@@ -5,6 +5,10 @@ use zbus::Result;
 use zbus::azync::Connection;
 #[cfg(not(feature = "azync"))]
 use zbus::Connection;
+#[cfg(feature = "azync")]
+use zbus::azync::Proxy;
+#[cfg(not(feature = "azync"))]
+use zbus::Proxy;
 
 use crate::{DEFAULT_DEST, generated::seat, types::{DbusPath}};
 
@@ -25,35 +29,51 @@ use crate::{DEFAULT_DEST, generated::seat, types::{DbusPath}};
 /// 
 /// assert!(manager.can_suspend().is_ok());
 /// ```
-pub struct SeatInterface<'a> {
-    _inner: seat::SeatProxy<'a>,
+pub struct SeatProxy<'a>(seat::SeatProxy<'a>);
+
+impl<'a> std::ops::Deref for SeatProxy<'a> {
+    type Target = Proxy<'a>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
 }
 
-impl<'a> SeatInterface<'a> {
+impl<'a> std::ops::DerefMut for SeatProxy<'a> {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
+
+impl<'a> std::convert::AsRef<Proxy<'a>> for SeatProxy<'a> {
+    fn as_ref(&self) -> &Proxy<'a> {
+        &*self
+    }
+}
+
+impl<'a> std::convert::AsMut<Proxy<'a>> for SeatProxy<'a> {
+    fn as_mut(&mut self) -> &mut Proxy<'a> {
+        &mut *self
+    }
+}
+
+impl<'a> SeatProxy<'a> {
     #[inline]
     pub fn new(connection: &Connection, path: &'a DbusPath) -> Result<Self> {
-        Ok(Self {
-            _inner: seat::SeatProxy::new_for(&connection, DEFAULT_DEST, path.path())?,
-        })
-    }
-
-    /// Borrow the underlying `SeatProxy` for use with zbus directly
-    #[inline]
-    pub fn get_proxy(&self) -> &seat::SeatProxy {
-        &self._inner
+        Ok(Self(seat::SeatProxy::new_for(&connection, DEFAULT_DEST, path.path())?))
     }
 
     /// Brings the session with the specified ID into the foreground if the
     /// session_id matches
     #[inline]
     pub fn activate_session(&self, session_id: &str) -> zbus::Result<()> {
-        self._inner.activate_session(session_id)
+        self.0.activate_session(session_id)
     }
 
     /// Switches to the session on the virtual terminal 
     #[inline]
     pub fn switch_to(&self, vtnr: u32) -> zbus::Result<()> {
-        self._inner.switch_to(vtnr)
+        self.0.switch_to(vtnr)
     }
 
     /// Switches to next session on the virtual terminal
@@ -61,7 +81,7 @@ impl<'a> SeatInterface<'a> {
     /// If there is no active session, switches to the first session.
     #[inline]
     pub fn switch_to_next(&self) -> zbus::Result<()> {
-        self._inner.switch_to_next()
+        self.0.switch_to_next()
     }
 
     /// Switches to previous session on the virtual terminal
@@ -69,76 +89,76 @@ impl<'a> SeatInterface<'a> {
     /// If there is no active session, switches to the first session.
     #[inline]
     pub fn switch_to_previous(&self) -> zbus::Result<()> {
-        self._inner.switch_to_previous()
+        self.0.switch_to_previous()
     }
 
     /// Nuke the seat
     #[inline]
     pub fn terminate(&self) -> zbus::Result<()> {
-        self._inner.terminate()
+        self.0.terminate()
     }
 
     /// Property: currently active session if there is one
     #[inline]
     pub fn get_active_session(&self) -> zbus::Result<DbusPath> {
-        self._inner.active_session()
+        self.0.active_session()
     }
 
     /// Property: the session is suitable for graphical logins
     #[inline]
     pub fn get_can_graphical(&self) -> zbus::Result<bool> {
-        self._inner.can_graphical()
+        self.0.can_graphical()
     }
 
     /// Property: the session is suitable for text logins
     #[inline]
     pub fn get_can_tty(&self) -> zbus::Result<bool> {
-        self._inner.can_tty()
+        self.0.can_tty()
     }
 
     /// Property: seat ID
     #[inline]
     pub fn get_id(&self) -> zbus::Result<String> {
-        self._inner.id()
+        self.0.id()
     }
 
     /// Property: Is the seat idle
     #[inline]
     pub  fn get_idle_hint(&self) -> zbus::Result<bool> {
-        self._inner.idle_hint()
+        self.0.idle_hint()
     }
 
     /// Property: timestamp of the last change of the idle hint boolean (realtime)
     #[inline]
     pub  fn get_idle_since_hint(&self) -> zbus::Result<Duration> {
-        self._inner.idle_since_hint().map(|t| Duration::from_micros(t))
+        self.0.idle_since_hint().map(|t| Duration::from_micros(t))
     }
 
     /// Property: timestamp of the last change of the idle hint boolean (walltime)
     #[inline]
     pub  fn get_idle_since_hint_monotonic(&self) -> zbus::Result<Duration> {
-        self._inner.idle_since_hint().map(|t| Duration::from_micros(t))
+        self.0.idle_since_hint().map(|t| Duration::from_micros(t))
     }
 
     // /// Property: sessions on this seat
     // #[inline]
     // pub fn get_sessions(&self) -> zbus::Result<Vec<DbusPath>> {
-    //     self._inner.sessions()
+    //     self.0.sessions()
     // }
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::ManagerInterface;
-    use crate::SeatInterface;
+    use crate::ManagerProxy;
+    use crate::SeatProxy;
     use zbus::Connection;
 
     #[test]
     fn timestamps() {
         let connection = Connection::new_system().unwrap();
-        let manager = ManagerInterface::new(&connection).unwrap();
+        let manager = ManagerProxy::new(&connection).unwrap();
         let seats = manager.list_seats().unwrap();
-        let seat = SeatInterface::new(&connection, &seats[0]).unwrap();
+        let seat = SeatProxy::new(&connection, &seats[0]).unwrap();
 
         assert!(seat.get_active_session().is_ok());
     }
@@ -146,9 +166,9 @@ mod tests {
     #[test]
     fn properties() {
         let connection = Connection::new_system().unwrap();
-        let manager = ManagerInterface::new(&connection).unwrap();
+        let manager = ManagerProxy::new(&connection).unwrap();
         let seats = manager.list_seats().unwrap();
-        let seat = SeatInterface::new(&connection, &seats[0]).unwrap();
+        let seat = SeatProxy::new(&connection, &seats[0]).unwrap();
 
         assert!(seat.get_active_session().is_ok());
         assert!(seat.get_can_graphical().is_ok());