+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+# [0.5.0] - 2021-03-15
+### Changed
+- Add deref, deref_mut, as_ref, as_mut to all proxy to coeerce to `Proxy`
+- Rename all `<Name>Interface` to `<Name>Proxy` as correct
+
 # [0.4.0] - 2021-03-15
 ### Changed
 - Begin using zbus 2.0